+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * After the next character sync the ability to sync action uses will be available.
 * If you have restricted import to trusted users only, you can now allow those users to sync their characters.
+* Parse things like Way of Astral Self Wisdom based martial arts attack.
 
 # 0.3.5